backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
requests>=2.32.0
pillow>=11.0.0
google-genai>=1.14.0
firebase-admin>=6.6.0
reportlab>=4.2.0
httpx>=0.28.0
qrcode[pil]>=8.0
pytest>=8.0.0
numpy>=1.26.0
aiofiles>=24.1.0
python-multipart>=0.0.18
